+++ src/main/java/onlyoffice/managers/auth/AuthContext.java
@@ -18,8 +18,9 @@
 
 package onlyoffice.managers.auth;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
 import java.io.IOException;
 
 public interface AuthContext {